Refactors all bae I/O into a single namespace.
